feat: add lang menu
