+++ src/app/core/components/footer/footer.component.html
@@ -9,13 +9,4 @@
       ><img class="footer__img" [src]="link.avatar" alt=""
     /></a>
   </div>
-
-  <mat-button-toggle-group name="fontStyle" aria-label="Font Style">
-    <mat-button-toggle
-      *ngFor="let lang of languages"
-      (change)="changeLanguage(lang)"
-      [checked]="lang === currentLanguage"
-      >{{ lang.toLocaleString().toLocaleUpperCase() }}</mat-button-toggle
-    >
-  </mat-button-toggle-group>
 </div>
