+++ src/app/core/components/footer/footer.component.scss
@@ -29,13 +29,3 @@
   margin: 0 4px;
   border-radius: 50%;
 }
-
-.mat-button-toggle-group {
-  min-width: 150px;
-  z-index: 20;
-}
-
-::ng-deep .mat-button-toggle-label-content {
-  line-height: 36px !important;
-  padding: 0 16px !important;
-}
