+++ src/app/core/components/header/header.component.html
@@ -16,5 +16,22 @@
       [active]="rla.isActive"
       >{{ link.label | translate }}</a
     >
+
+    <div>
+      <button mat-icon-button color="primary" [matMenuTriggerFor]="menu">
+        <mat-icon>language</mat-icon>
+      </button>
+
+      <mat-menu #menu="matMenu">
+        <button
+          [disabled]="lang === currentLanguage"
+          *ngFor="let lang of languages"
+          mat-menu-item
+          (click)="changeLanguage(lang)"
+        >
+          {{ lang.toLocaleString().toLocaleUpperCase() }}
+        </button>
+      </mat-menu>
+    </div>
   </nav>
 </div>
